backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
python-multipart
pymongo
python-jose[cryptography]
passlib
bcrypt==3.2.2
gunicorn